% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{match_find}
\alias{match_find}
\title{Fuzzy matching of a text string}
\usage{
match_find(text, pattern, loc)
}
\arguments{
\item{text}{The text to search.}

\item{pattern}{The pattern to search for.}

\item{loc}{The location to search around.}
}
\value{
Best match index or -1.
}
\description{
Locate the best instance of \code{pattern} in the \code{text} near \code{loc} using the
Bitap algorithm.Returns \code{-1} if no match found. Assumes R's typical 1-based indexing for \code{loc}
and the returned value.

This algorithm makes use of the \code{match_distance} and \code{match_threshold} options to determine
the match.

Candidate matches are scored based on a) the number of spelling differences between the pattern
and the text and b) the distance between the candidate match and the expected location. The
\code{match_distance} option determines the relative importance of these two metrics.
}
