% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{patch_make}
\alias{patch_make}
\alias{patch_apply}
\title{Create and apply patches to a text string}
\usage{
patch_make(x, y)

patch_apply(x, patch)
}
\arguments{
\item{x}{The source string}

\item{y}{The destination string}

\item{patch}{A string representation of the patch(es).}
}
\value{
\code{patch_make()} returns a string representation of the patch(es).

\code{patch_apply()} returns the patched version of the string x, the \code{matches} attribute
contains logical values indicating which patches were successfully applied.
}
\description{
\code{patch_make()} computes a list of patches to turn the source string (\code{x}) into the
destination string (\code{y}).

\code{patch_apply()} applies the \code{patch} to \code{x}.
}
